Update dependencies in poetry.lock to latest versions and add new packages
This commit updates several package versions in the `poetry.lock` file, including `anyio` to 4.9.0, `attrs` to 25.3.0, `bokeh` to 3.7.3, and `boto3` to 1.38.32. Additionally, it introduces new packages such as `async-timeout` (5.0.1), `blinker` (1.9.0), and `zipp` (3.22.0). The changes also include adjustments to dependency specifications and extra requirements for some packages, enhancing compatibility and functionality across the project.